add wack skinchanger + misc fixes

+++ cs2cheat/src/sdk/source-sdk/classes/entity/chandle.hpp
@@ -16,5 +16,5 @@ class CHandle {
         return (T*)(Get());
     }
 
-    uintptr_t m_Index;
+    uint32_t m_Index;
 };
